+ Adding new package and class for exercise formulas which are partly realized.

+++ src/main/java/org/example/geometry/CalculateY.java
@@ -1,5 +1,5 @@
 package org.example.geometry;
 
 public interface CalculateY {
-    float calculateYByFormula(float x);
+    double calculateYByFormula(double x);
 }
